@@ -22,16 +22,13 @@ public static class Extensions
     /// Does not give precision of 7 if value of sender does not have 7 places while <see cref="GetMilliseconds7"/> does
     /// </returns>
     public static int GetMilliseconds(this DateTime sender) 
-        => Convert.ToInt32(sender.TimeOfDay.Milliseconds.ToString() + sender.TimeOfDay.Microseconds.ToString());
+        => Convert.ToInt32(sender.TimeOfDay.Milliseconds.ToString() + sender.TimeOfDay.Microseconds);
 
     /// <summary>
     /// Ensures precision is 7
     /// </summary>
     public static int GetMilliseconds7(this DateTime sender) 
-        => Convert.ToInt32((sender.TimeOfDay.Milliseconds.ToString() + sender.TimeOfDay.Microseconds.ToString())
+        => Convert.ToInt32((sender.TimeOfDay.Milliseconds.ToString() + sender.TimeOfDay.Microseconds)
             .PadRight(7, '0'));
 
-    // think this might work, no as it does addition
-    //public static int GetMilliseconds(this DateTime sender) 
-    //    => Convert.ToInt32(sender.TimeOfDay.Milliseconds + sender.TimeOfDay.Microseconds);
 }
